Part: move ViewProviderPrimitive to its own source files
