+++ src/Mod/Part/Gui/ViewProvider.h
@@ -61,26 +61,6 @@ protected:
                     std::vector<App::Color>& colors);
 };
 
-class PartGuiExport ViewProviderPrimitive : public ViewProviderPart
-{
-    PROPERTY_HEADER(PartGui::ViewProviderPrimitive);
-
-public:
-    /// constructor
-    ViewProviderPrimitive();
-    /// destructor
-    virtual ~ViewProviderPrimitive();
-
-    void setupContextMenu(QMenu*, QObject*, const char*);
-
-protected:
-    bool setEdit(int ModNum);
-    void unsetEdit(int ModNum);
-
-private:
-    void startDefaultEditMode();
-};
-
 } // namespace PartGui
